more changes

+++ pages/users/index.jsx
@@ -1,10 +1,12 @@
 import { useState, useEffect } from "react";
 
-import { Link, Spinner } from "components";
-import { Layout } from "components/users";
+import { Spinner } from "components";
+
 import { userService } from "services";
 
 import { useRouter } from "next/router";
+import Link from "next/link";
+import Layout from "../../components/account/Layout";
 
 export default Index;
 
@@ -32,48 +34,57 @@ function Index() {
 
   if (userService.userValue.username === "admin") {
     return (
-      <Layout custom="bg-red-900">
+      <Layout custom="mt-20">
         <h1>Users</h1>
-        <Link href="/users/add" className="btn btn-sm btn-success mb-2">
+        <Link
+          href="/users/add"
+          legacyBehavior
+          className="btn btn-sm btn-success mb-2"
+        >
           Add User
         </Link>
-        <table className="table table-striped">
+        <table className="table table-striped text-white">
           <thead>
             <tr>
               <th style={{ width: "30%" }}>First Name</th>
               <th style={{ width: "30%" }}>Last Name</th>
               <th style={{ width: "30%" }}>Username</th>
-              <th style={{ width: "10%" }}></th>
+              <th style={{ width: "10%" }} />
             </tr>
           </thead>
           <tbody>
             {users &&
-              users.map((user) => (
-                <tr key={user.id}>
-                  <td>{user.firstName}</td>
-                  <td>{user.lastName}</td>
-                  <td>{user.username}</td>
-                  <td style={{ whiteSpace: "nowrap" }}>
-                    <Link
-                      href={`/users/edit/${user.id}`}
-                      className="btn btn-sm btn-primary mr-1"
-                    >
-                      Edit
-                    </Link>
-                    <button
-                      onClick={() => deleteUser(user.id)}
-                      className="btn btn-sm btn-danger btn-delete-user"
-                      disabled={user.isDeleting}
-                    >
-                      {user.isDeleting ? (
-                        <span className="spinner-border spinner-border-sm"></span>
-                      ) : (
-                        <span>Delete</span>
-                      )}
-                    </button>
-                  </td>
-                </tr>
-              ))}
+              users.map((user) => {
+                if (user.username !== "admin") {
+                  return (
+                    <tr key={user.id}>
+                      <td>{user.firstName}</td>
+                      <td>{user.lastName}</td>
+                      <td>{user.username}</td>
+                      <td style={{ whiteSpace: "nowrap" }}>
+                        <Link
+                          href={`/users/edit/${user.id}`}
+                          legacyBehavior
+                          className="btn btn-sm btn-primary mr-1"
+                        >
+                          Edit
+                        </Link>
+                        <div
+                          onClick={() => deleteUser(user.id)}
+                          className="btn btn-sm btn-danger btn-delete-user ml-4"
+                          disabled={user.isDeleting}
+                        >
+                          {user.isDeleting ? (
+                            <span className="spinner-border spinner-border-sm" />
+                          ) : (
+                            <span>Delete</span>
+                          )}
+                        </div>
+                      </td>
+                    </tr>
+                  );
+                }
+              })}
             {!users && (
               <tr>
                 <td colSpan="4">
